Make union value trees lazily generated.
This makes union value trees lazily generated. (Tuple-based unions are converted over in the next diff.)

The main insight here is that if a copy of the strategy is kept in the value tree, values can be generated
at shrinking time instead of in advance. This leads to significant speedups in generation times for large
enums, as will be seen in upcoming benchmarks.