Simplifying copyright
"Copyright Headers" of CONTRIBUTING.md originally suggested to add your own
name for copyright. However, we support both MIT and Apache licenses which
allow anyone to copy and modify this source for their own use. For
simplification then this change asks everyone to copyright code with the name
"The proptest developers".