Fix bounds generation for generics in derive(Arbitrary) (#511)
The implementation of UseTracker expects that iteration over items of
used_map gives items in insertion order. However, the order of BTreeSet
is based on Ord, not insertion.

I use a Vec of tuples. An alternative would be indexmap crate, but since
the maps are supposed to be very small, it is probably not worthy. There
already was a mention about the crate in the comments worrying about
compile times.

Co-authored-by: Petr Nevyhoštěný <petr.nevyhosteny@gmail.com>