fix state-machine shrinking on <= 1 transitions
normally we start deleting transitions earlier than the last failing
transition as our first pass at shrinking. we cannot delete the
last seen/failing transition because we know it is causal in the
failure. however the code would perform this deletion step even when
seen transitions <= 1 and then default to Deletion(0) which at min
is inconsistent with expectations.

it turns out this issue always existed but for some reason rand v0.8
rarely would test on transition seen states < 2 whereas rand v0.9
does it often enough that over 100 iterations tests are likely to
fail. thus during our rand upgrade we ran into this behavior in
our tests.

the fix is on 1 transition seen we set shrinking to Transition(0)
and on 0 transitions seen we set shrinking to InitialState which
is congruent with expectations.